+++ core/include/snobol/snobol.h
@@ -17,9 +17,9 @@
 
 /* Version macros */
 #define SNOBOL_VERSION_MAJOR 0
-#define SNOBOL_VERSION_MINOR 2
+#define SNOBOL_VERSION_MINOR 3
 #define SNOBOL_VERSION_PATCH 0
-#define SNOBOL_VERSION_STRING "0.2.0"
+#define SNOBOL_VERSION_STRING "0.3.0"
 
 /**
  * Get library version at runtime